==> Cxcl12_ko/1_cxcl12_ko_read_data_and_label.rmd <==
---
title: "scSEQ analysis of CXCL12+ endothelium"
author: Nick Negretti
date: 11/02/21
output: rmarkdown::github_document
---

## Load libraries and helper functions
```{r, results="hide", message = FALSE}
setwd("~/code/LungVascularDevelopment_github")
renv::activate("~/code/LungVascularDevelopment_github")

source("./helper_functions/libraries.R")

opts_knit$set(root.dir = getwd())

source("./helper_functions/trajectory.R")
source("./helper_functions/cluster_scanpy.R")
source("./helper_functions/colors.R")
source("./helper_functions/brackets.R")
source("./helper_functions/heatmaps.R")
set.seed(42) # For reproducability

N_WORKERS <- 4
options(future.globals.maxSize=6 *1024*1024^2) # First number is shared object size in GB
plan("multiprocess", workers = N_WORKERS)
```

```{r}
runSoupX <- function(raw_counts, filtered_counts, mat_out_path, project_name){
  in_soup <- SoupChannel(tod = raw_counts[rownames(raw_counts) != "Gm42418",],
                         toc = filtered_counts[rownames(filtered_counts) != "Gm42418",],
                         metaData = NULL, soupRange = c(0, 10), keepDroplets = FALSE) # Gm42418 is a rRNA contaminant - remove it now
  sr_obj <- CreateSeuratObject(in_soup$toc)
  # Fast preliminary clustering and normalization for soupx run
  sr_obj <- NormalizeData(sr_obj)
  sr_obj <- FindVariableFeatures(sr_obj)
  sr_obj <- ScaleData(sr_obj)
  sr_obj <- RunPCA(sr_obj,pcs.compute=30)
  sr_obj <- RunUMAP(sr_obj,dims=1:15)
  sr_obj <- FindNeighbors(sr_obj,dims=1:15)
  sr_obj <- FindClusters(sr_obj,resolution=1.2)
  sr_DF <-  as.data.frame(sr_obj@reductions$umap@cell.embeddings)
  colnames(sr_DF) <- c('RD1','RD2')
  sr_DF$Cluster <- factor(sr_obj@meta.data[rownames(sr_DF),'RNA_snn_res.1.2'])

  #Specify background RNA genes
  background_RNA_genes <- c("Dcn", "Bgn", "Aspn", "Ecm2", "Fos", "Hbb-bs", "Hbb-bt", "Hba-a1", "Hba-a2", "Lyz1", "Lyz2", "Mgp", "Postn", "Scgb1a1")

  #Specify background RNA genes for epithelial specific cell sort
  est_expression <- estimateNonExpressingCells(in_soup,
                                               list(SLRP = background_RNA_genes),
                                               clusters = setNames(sr_DF$Cluster, rownames(sr_DF)),
                                               FDR = 0.05,
                                               maximumContamination = 1.5)

  ## calculate contamination and adjust the counts.
  in_soup <- setClusters(in_soup, sr_DF$Cluster)
  in_soup <- calculateContaminationFraction(in_soup, list(SLRP = background_RNA_genes), useToEst = est_expression, forceAccept = TRUE)
  head(in_soup$metaData)

  ## adjust the counts based on contamination fraction
  ## Return dgCMatrix
  #return(adjustCounts(in_soup, roundToInt = TRUE))
  adj_counts <- adjustCounts(in_soup)
  #Matrix::writeMM(adj_counts, file = paste0(mat_out_path))
  DropletUtils::write10xCounts(mat_out_path, adj_counts, overwrite = TRUE)
  out_object <- CreateSeuratObject(counts = adj_counts, project = project_name)
  out_object[["RAW_COUNTS"]] <- CreateAssayObject(counts = filtered_counts[rownames(out_object),])
  DefaultAssay(out_object) <- "RNA"
  return(out_object)
}
```

```{r}
pct_mt_high <- 10
pct_mt_low <- 0.5
min_nfeature <- 700
# Change to %<-% to make future objects - fails on Mac because of temporary file bug
het <- {runSoupX(Read10X("./data/Cxcl12_KO/E18_het/raw_feature_bc_matrix"),
                      Read10X("./data/Cxcl12_KO/E18_het/filtered_feature_bc_matrix"),
                      "./data/Cxcl12_KO/Cxcl12_het_soupx_counts/",
                      "het") %>%
        PercentageFeatureSet(., pattern = "^mt-", col.name = "percent.mt") %>%
        subset(., subset = percent.mt < pct_mt_high & percent.mt > pct_mt_low & nFeature_RNA > min_nfeature)}

ko <- {runSoupX(Read10X("./data/Cxcl12_KO/E18_ko/raw_feature_bc_matrix"),
                      Read10X("./data/Cxcl12_KO/E18_ko/filtered_feature_bc_matrix"),
                      "./data/Cxcl12_KO/Cxcl12_ko_soupx_counts/",
                      "ko") %>%
        PercentageFeatureSet(., pattern = "^mt-", col.name = "percent.mt") %>%
        subset(., subset = percent.mt < pct_mt_high & percent.mt > pct_mt_low & nFeature_RNA > min_nfeature)}
```


## Load counts and merge

```{r}
# Cache the results - reload from stored object if alreay complete
filename <- "./data/cxcl12_ko_merged_notransform.rds"
if (!file.exists(filename)) {
  pct_mt_high <- 10
  pct_mt_low <- 0.5
  min_nfeature <- 700
  # Change to %<-% to make future objects - fails on Mac because of temporary file bug
  het <- {runSoupX(Read10X("./data/Cxcl12_KO/E18_het/raw_feature_bc_matrix"),
                   Read10X("./data/Cxcl12_KO/E18_het/filtered_feature_bc_matrix"),
                   "./data/Cxcl12_KO/Cxcl12_het_soupx_counts/",
                   "het") %>%
          PercentageFeatureSet(., pattern = "^mt-", col.name = "percent.mt") %>%
          subset(., subset = percent.mt < pct_mt_high & percent.mt > pct_mt_low & nFeature_RNA > min_nfeature)}

  ko <- {runSoupX(Read10X("./data/Cxcl12_KO/E18_ko/raw_feature_bc_matrix"),
                  Read10X("./data/Cxcl12_KO/E18_ko/filtered_feature_bc_matrix"),
                  "./data/seq/Cxcl12_KO/Cxcl12_KO_soupx_counts/",
                  "ko") %>%
          PercentageFeatureSet(., pattern = "^mt-", col.name = "percent.mt") %>%
          subset(., subset = percent.mt < pct_mt_high & percent.mt > pct_mt_low & nFeature_RNA > min_nfeature)}
  # This is done on one line to reduce the number of redundant objects stored in memory
  # Project sets 'orig.ident' in the resulting object
  obj <- Reduce(function(x,y){merge(x,y)}, list(het, ko))

  obj <- PercentageFeatureSet(obj, pattern = "^mt-", col.name = "percent.mt")

  # Filter cells to remove those with high mtRNA and low gene counts
  obj <- subset(obj, subset = percent.mt < 10 & percent.mt >0.5 & nFeature_RNA >700)
  saveRDS(obj, filename, compress = "bzip2")
} else {
  obj <- readRDS(filename)
}
```



## Check number of cells at each condition
```{r}
table(obj$orig.ident)
```

# Run SCTransform
```{r}
plan("multiprocess", workers = N_WORKERS)
filename <- "./data/cxcl12_ko_merged_sct.rds"
if (!file.exists(filename)) {
  obj <- SCTransform(obj, variable.features.n = 2000, conserve.memory = TRUE, batch_var = "orig.ident", vars.to.regress = c("percent.mt"),
                     method = "glmGamPoi")
  saveRDS(obj, filename, compress = "bzip2")
} else {
  obj <- readRDS(filename)
}
```

## Preliminary clustering and UMAP
```{r, fig.width = 12}
set.seed(42)
obj <- cluster_pca_umap(obj, k_param = 15, dims_umap = 1:25, dims_neighbors = 1:25, cluster_res = 1)
cluster_plot <- DimPlot(obj, label = TRUE) + umap_theme() + theme(aspect.ratio = 1)
timepoint_plot <- DimPlot(obj, group.by = "orig.ident") + umap_theme() + theme(aspect.ratio = 1)

cluster_plot + timepoint_plot + plot_layout(nrow = 1)
```

```{r}
VlnPlot(obj, features = "Cxcl12", group.by = "seurat_clusters", split.by = "orig.ident")
```

```{r}
marker_genes <- c("Hbb-bs", "Epcam", "Foxj1", "Scgb1a1", "Scgb3a2", "Abca3", "Hopx", "Col1a1", "Dcn", "Lum", "Acta2", "Wnt2", "Wnt5a", "Lgr6", "Pdgfra", "Pdgfrb", "Cspg4", "Wt1", "Pecam1", "Ccl21a", "Vwf", "Nrg1", "Plvap", "Car4", "Mki67", "Tnnt2", "Myl7", "Ascl1", "Mal", "Ptprc")
DotPlot(obj, features = marker_genes) + theme(axis.text.x = element_text(angle = 45, hjust=1))
```

```{r}
obj_relabel <- obj
obj_relabel$bulk_celltype <- as.factor(Idents(obj))

idxes <- levels(obj_relabel$bulk_celltype)
labels <- rep("", length(levels(obj_relabel$bulk_celltype)))
labels[idxes %in% c(8, 9, 15, 22)] <- "Immune"
labels[idxes %in% c(14, 17, 18, 20)] <- "Epithelium"
labels[idxes %in% c(0, 1, 3, 5, 6, 12, 13, 19, 21, 24, 25, 28)] <- "Mesenchyme"
labels[idxes %in% c(2, 4, 7, 10, 11, 16)] <- "Endothelium"
labels[idxes %in% c(26)] <- "Neuronal"  # High Mal, Foxd3 expression / embryonic spinal cord?
labels[idxes %in% c(23, 27)] <- "Doublet"
labels

levels(obj_relabel$bulk_celltype) <- labels
obj_relabel_sub <- subset(obj_relabel, bulk_celltype %in% c("Neuronal", "Doublet"), invert = TRUE)

```


```{r}
DimPlot(obj_relabel_sub, group.by = "bulk_celltype") + umap_theme() + theme(aspect.ratio = 1)
```

```{r}
marker_genes <- c("Mdk", # Early Epithelium
                  "Cdkn1a", "Krt8", # Transitional
                  "Hopx", "Aqp5", #AT1
                  "Sftpa1", "Sftpc", # AT2
                  "Foxj1", "Dynlrb2", # Ciliated
                  "Scgb1a1", "Scgb3a2", # Secretory
                  "Ascl1", "Scg5", #Neuroendocrine

                  "Cxcl12", "Pcsk5", # Art maEC
                  "Gkn3", # Proximal Arterial
                  "Gja5", # Distal Arterial
                  "Vegfc", "Prss23", # Ven maEC
                  "Mki67", "Top2a", # Prolif gCap
                  "Gpihbp1", "Kit", # gCap
                  "Car4", "Kdr", # aCap
                  "Flt4", "Ccl21a", # Lymphatic

                  "Wnt2", "Macf1", # Wnt2 fibroblasts 0
                  "Tgfbi", "Wnt5a", # Myofibroblasts 1
                  "Dcn", "Col1a1", # Adventitial fibroblast 2
                  "Cspg4", "Cox4i2", # Pericyte 4
                  "Wt1", "Upk3b", # Mesothelium 6
                  "Eln", "Acta2", # Smooth muscle 7
                  "Tnnt2", "Actc1", # Cardiomyocyte
                  "Ptprc"
)
```

For doublet removal, split in to broad celltypes - analyze separately, then merge back together. This will give better resolution.

```{r}
cxcl12_ko_epi_prelim <- subset(obj_relabel_sub, bulk_celltype == "Epithelium")
cxcl12_ko_endo_prelim <- subset(obj_relabel_sub, bulk_celltype == "Endothelium")
cxcl12_ko_meso_prelim <- subset(obj_relabel_sub, bulk_celltype == "Mesenchyme")
cxcl12_ko_immune_prelim <- subset(obj_relabel_sub, bulk_celltype == "Immune")
```


```{r}
cxcl12_ko_epi_prelim_sub <- subset(cxcl12_ko_epi_prelim, idents = c(15, 19), invert = TRUE)
cxcl12_ko_epi_prelim_sub <- cluster_pca_umap(cxcl12_ko_epi_prelim_sub, k_param = 10, dims_umap = 1:15, dims_neighbors = 1:15, cluster_res = 0.4)
cluster_plot <- DimPlot(cxcl12_ko_epi_prelim_sub, label = TRUE) + umap_theme() + theme(aspect.ratio = 1)
celltype_plot <- DimPlot(cxcl12_ko_epi_prelim_sub, group.by = "bulk_celltype") + umap_theme() + theme(aspect.ratio = 1)
genotype_plot <- DimPlot(cxcl12_ko_epi_prelim_sub, group.by = "orig.ident") + umap_theme() + theme(aspect.ratio = 1)
cluster_plot + genotype_plot
```


```{r}
DotPlot(cxcl12_ko_epi_prelim_sub, features = marker_genes) + theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

```{r}
cxcl12_ko_epi_prelim_relabel <- cxcl12_ko_epi_prelim_sub
cxcl12_ko_epi_prelim_relabel$celltype <- as.factor(Idents(cxcl12_ko_epi_prelim_sub))

idxes <- levels(cxcl12_ko_epi_prelim_relabel$celltype)
labels <- rep("", length(levels(cxcl12_ko_epi_prelim_relabel$celltype)))
labels[idxes %in% c(1, 6)] <- "AT1"
labels[idxes %in% c(2, 3, 4)] <- "AT2"
labels[idxes %in% c(5, 7, 8)] <- "Ciliated"
labels[idxes %in% c(0)] <- "Secretory"
labels

levels(cxcl12_ko_epi_prelim_relabel$celltype) <- labels

```


```{r}
cluster_plot <- DimPlot(cxcl12_ko_epi_prelim_relabel, label = TRUE) + umap_theme() + theme(aspect.ratio = 1)
celltype_plot <- DimPlot(cxcl12_ko_epi_prelim_relabel, group.by = "celltype") + umap_theme() + theme(aspect.ratio = 1)
genotype_plot <- DimPlot(cxcl12_ko_epi_prelim_relabel, group.by = "orig.ident") + umap_theme() + theme(aspect.ratio = 1)
cluster_plot + celltype_plot + genotype_plot
```

```{r}
markers <- FindMarkers(cxcl12_ko_epi_prelim_relabel, 6, 1)
markers$gene <- rownames(markers)  # Bugfix for display
markers
```

# Label endothelial subset

```{r}
# low k-param to get small clusters of doublets into their own groups
cxcl12_ko_endo_prelim <- cluster_pca_umap(cxcl12_ko_endo_prelim, k_param = 10, dims_umap = 1:25, dims_neighbors = 1:25, cluster_res = 1.0)
cluster_plot <- DimPlot(cxcl12_ko_endo_prelim, label = TRUE) + umap_theme() + theme(aspect.ratio = 1)
celltype_plot <- DimPlot(cxcl12_ko_endo_prelim, group.by = "bulk_celltype") + umap_theme() + theme(aspect.ratio = 1)
genotype_plot <- DimPlot(cxcl12_ko_endo_prelim, group.by = "orig.ident") + umap_theme() + theme(aspect.ratio = 1)
cluster_plot + genotype_plot
```

```{r}
DotPlot(cxcl12_ko_endo_prelim, features = marker_genes) + theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

```{r}
markers <- FindMarkers(cxcl12_ko_endo_prelim, 6, NULL)
markers$gene <- rownames(markers)
markers
```

```{r}
FeaturePlot(cxcl12_ko_endo_prelim, "Gkn3")
```

```{r}
cxcl12_ko_endo_prelim_relabel <- cxcl12_ko_endo_prelim
cxcl12_ko_endo_prelim_relabel$celltype <- as.factor(Idents(cxcl12_ko_endo_prelim))

idxes <- levels(cxcl12_ko_endo_prelim_relabel$celltype)
labels <- rep("", length(levels(cxcl12_ko_endo_prelim_relabel$celltype)))
labels[idxes %in% c(12)] <- "Proximal Arterial Endothelium"
labels[idxes %in% c(1, 7, 9)] <- "Distal Arterial Endothelium"
labels[idxes %in% c(13)] <- "Venous Endothelium"
labels[idxes %in% c(0, 2, 3, 4, 11)] <- "General Capillaries (gCap)"
labels[idxes %in% c(5, 10)] <- "Prolif. Capillaries (gCap)"
labels[idxes %in% c(8)] <- "Car4+ Capillaries"
labels[idxes %in% c(14)] <- "Lymphatic"
labels[idxes %in% c(6)] <- "Unknown / Other"

labels

levels(cxcl12_ko_endo_prelim_relabel$celltype) <- labels

```
```{r}
cluster_plot <- DimPlot(cxcl12_ko_endo_prelim_relabel, label = TRUE) + umap_theme() + theme(aspect.ratio = 1)
celltype_plot <- DimPlot(cxcl12_ko_endo_prelim_relabel, group.by = "celltype") + umap_theme() + theme(aspect.ratio = 1)
genotype_plot <- DimPlot(cxcl12_ko_endo_prelim_relabel, group.by = "orig.ident") + umap_theme() + theme(aspect.ratio = 1)
cluster_plot + celltype_plot + genotype_plot
```

# Mesenchymal cells
```{r}
# low k-param to get small clusters of doublets into their own groups
cxcl12_ko_meso_prelim <- cluster_pca_umap(cxcl12_ko_meso_prelim, k_param = 10, dims_umap = 1:25, dims_neighbors = 1:25, cluster_res = 1.0)
cluster_plot <- DimPlot(cxcl12_ko_meso_prelim, label = TRUE) + umap_theme() + theme(aspect.ratio = 1)
celltype_plot <- DimPlot(cxcl12_ko_meso_prelim, group.by = "bulk_celltype") + umap_theme() + theme(aspect.ratio = 1)
genotype_plot <- DimPlot(cxcl12_ko_meso_prelim, group.by = "orig.ident") + umap_theme() + theme(aspect.ratio = 1)
cluster_plot + genotype_plot
```

```{r}
markers <- FindMarkers(cxcl12_ko_meso_prelim, 18, NULL)
markers$gene <- rownames(markers)
markers
```

```{r}
DotPlot(cxcl12_ko_meso_prelim, features = marker_genes) + theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

```{r}
# 16 is a doublet, 18 appears to be a few skin cells
cxcl12_ko_meso_prelim_sub <- subset(cxcl12_ko_meso_prelim, idents = c(16, 18), invert = TRUE)
cxcl12_ko_meso_prelim_sub <- cluster_pca_umap(cxcl12_ko_meso_prelim_sub, k_param = 10, dims_umap = 1:20, dims_neighbors = 1:20, cluster_res = 0.8)
cluster_plot <- DimPlot(cxcl12_ko_meso_prelim_sub, label = TRUE) + umap_theme() + theme(aspect.ratio = 1)
celltype_plot <- DimPlot(cxcl12_ko_meso_prelim_sub, group.by = "bulk_celltype") + umap_theme() + theme(aspect.ratio = 1)
genotype_plot <- DimPlot(cxcl12_ko_meso_prelim_sub, group.by = "orig.ident") + umap_theme() + theme(aspect.ratio = 1)
cluster_plot + genotype_plot
```
```{r}
DotPlot(cxcl12_ko_meso_prelim_sub, features = marker_genes) + theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

```{r}
markers <- FindMarkers(cxcl12_ko_meso_prelim_sub, 10, NULL)
markers$gene <- rownames(markers)
markers
```

```{r}
FeaturePlot(cxcl12_ko_meso_prelim_sub, "Axin2")
```

```{r}
cxcl12_ko_meso_prelim_relabel <- cxcl12_ko_meso_prelim_sub
cxcl12_ko_meso_prelim_relabel$celltype <- as.factor(Idents(cxcl12_ko_meso_prelim_relabel))

idxes <- levels(cxcl12_ko_meso_prelim_relabel$celltype)
labels <- rep("", length(levels(cxcl12_ko_meso_prelim_relabel$celltype)))
labels[idxes %in% c(0, 3, 9)] <- "Wnt2+ Fibroblast"
labels[idxes %in% c(1, 2, 4, 5, 7)] <- "Myofibroblast"
labels[idxes %in% c(12)] <- "Smooth Muscle"
labels[idxes %in% c(15)] <- "Adventitial Fibroblast"
labels[idxes %in% c(11, 13)] <- "Pericyte"
labels[idxes %in% c(16)] <- "Mesothelium"
labels[idxes %in% c(10)] <- "Unknown (Ptn+ Prss35+)"
labels[idxes %in% c(6, 8)] <- "Alveolar FB"
labels[idxes %in% c(14)] <- "Progenitors"

labels

levels(cxcl12_ko_meso_prelim_relabel$celltype) <- labels

```


```{r}
# MANC: Pdgfra+ Wnt2- Wnt5a- Tgfbi- Wif1+(partial) Mmp2+
cluster_plot <- DimPlot(cxcl12_ko_meso_prelim_relabel, label = TRUE) + umap_theme() + theme(aspect.ratio = 1)
celltype_plot <- DimPlot(cxcl12_ko_meso_prelim_relabel, group.by = "celltype") + umap_theme() + theme(aspect.ratio = 1)
genotype_plot <- DimPlot(cxcl12_ko_meso_prelim_relabel, group.by = "orig.ident") + umap_theme() + theme(aspect.ratio = 1)
cluster_plot + celltype_plot + genotype_plot
```


```{r}
epi_cluster_plot <- ggplot() + umap_theme() +
        ggrastr::rasterise(geom_point(data = cxcl12_ko_epi_prelim_relabel, aes(x = x, y = y, color = celltype), size = .05), dpi = 600) +
        scale_colour_manual(name = "Celltype", values = color_category_20,
                            guide = guide_legend(override.aes = list(size=3), order = 1)) +
        theme(aspect.ratio=1,
              legend.text=element_text(size=12),
              legend.title=element_text(size=12),
              panel.background = element_rect(fill = "transparent", colour = "black"),
              plot.background = element_rect(color = "transparent", fill = "transparent")) +
        ylab("UMAP2") + xlab("UMAP1")

epi_genotype_plot <- ggplot() + umap_theme() +
        ggrastr::rasterise(geom_point(data = cxcl12_ko_epi_prelim_relabel, aes(x = x, y = y, color = orig.ident), size = .05), dpi = 600) +
        scale_colour_manual(name = "Genotype", values = color_scanpy_default,
                            guide = guide_legend(override.aes = list(size=3), order = 1)) +
        theme(aspect.ratio=1,
              legend.text=element_text(size=12),
              legend.title=element_text(size=12),
              panel.background = element_rect(fill = "transparent", colour = "black"),
              plot.background = element_rect(color = "transparent", fill = "transparent")) +
        ylab("UMAP2") + xlab("UMAP1")

endo_cluster_plot <- ggplot() + umap_theme() +
        ggrastr::rasterise(geom_point(data = cxcl12_ko_endo_prelim_relabel, aes(x = x, y = y, color = celltype), size = .05), dpi = 600) +
        scale_colour_manual(name = "Celltype", values = color_category_20,
                            guide = guide_legend(override.aes = list(size=3), order = 1)) +
        theme(aspect.ratio=1,
              legend.text=element_text(size=12),
              legend.title=element_text(size=12),
              panel.background = element_rect(fill = "transparent", colour = "black"),
              plot.background = element_rect(color = "transparent", fill = "transparent")) +
        ylab("UMAP2") + xlab("UMAP1")

endo_genotype_plot <- ggplot() + umap_theme() +
        ggrastr::rasterise(geom_point(data = cxcl12_ko_endo_prelim_relabel, aes(x = x, y = y, color = orig.ident), size = .05), dpi = 600) +
        scale_colour_manual(name = "Genotype", values = color_scanpy_default,
                            guide = guide_legend(override.aes = list(size=3), order = 1)) +
        theme(aspect.ratio=1,
              legend.text=element_text(size=12),
              legend.title=element_text(size=12),
              panel.background = element_rect(fill = "transparent", colour = "black"),
              plot.background = element_rect(color = "transparent", fill = "transparent")) +
        ylab("UMAP2") + xlab("UMAP1")

meso_cluster_plot <- ggplot() + umap_theme() +
        ggrastr::rasterise(geom_point(data = cxcl12_ko_meso_prelim_relabel, aes(x = x, y = y, color = celltype), size = .05), dpi = 600) +
        scale_colour_manual(name = "Celltype", values = color_category_20,
                            guide = guide_legend(override.aes = list(size=3), order = 1)) +
        theme(aspect.ratio=1,
              legend.text=element_text(size=12),
              legend.title=element_text(size=12),
              panel.background = element_rect(fill = "transparent", colour = "black"),
              plot.background = element_rect(color = "transparent", fill = "transparent")) +
        ylab("UMAP2") + xlab("UMAP1")

meso_genotype_plot <- ggplot() + umap_theme() +
        ggrastr::rasterise(geom_point(data = cxcl12_ko_meso_prelim_relabel, aes(x = x, y = y, color = orig.ident), size = .05), dpi = 600) +
        scale_colour_manual(name = "Genotype", values = color_scanpy_default,
                            guide = guide_legend(override.aes = list(size=3), order = 1)) +
        theme(aspect.ratio=1,
              legend.text=element_text(size=12),
              legend.title=element_text(size=12),
              panel.background = element_rect(fill = "transparent", colour = "black"),
              plot.background = element_rect(color = "transparent", fill = "transparent")) +
        ylab("UMAP2") + xlab("UMAP1")


```


```{r}
ggsave("./data/figures/cxcl12_prelim_umap_epi.pdf", epi_cluster_plot + epi_genotype_plot + plot_layout(ncol = 2) &
        plot_annotation(theme = theme(plot.background = element_rect(color = "transparent", fill = "transparent"))),
       bg = "transparent")
ggsave("./data/figures/cxcl12_prelim_umap_endo.pdf", endo_cluster_plot + endo_genotype_plot + plot_layout(ncol = 2) &
        plot_annotation(theme = theme(plot.background = element_rect(color = "transparent", fill = "transparent"))),
       bg = "transparent")
ggsave("./data/figures/cxcl12_prelim_umap_meso.pdf", meso_cluster_plot + meso_genotype_plot + plot_layout(ncol = 2) &
        plot_annotation(theme = theme(plot.background = element_rect(color = "transparent", fill = "transparent"))),
       bg = "transparent")
```


# Immune labeling

```{r}
# low k-param to get small clusters of doublets into their own groups
cxcl12_ko_immune_prelim <- cluster_pca_umap(cxcl12_ko_immune_prelim, k_param = 10, dims_umap = 1:20, dims_neighbors = 1:20, cluster_res = 0.4)
cluster_plot <- DimPlot(cxcl12_ko_immune_prelim, label = TRUE) + umap_theme() + theme(aspect.ratio = 1)
celltype_plot <- DimPlot(cxcl12_ko_immune_prelim, group.by = "bulk_celltype") + umap_theme() + theme(aspect.ratio = 1)
genotype_plot <- DimPlot(cxcl12_ko_immune_prelim, group.by = "orig.ident") + umap_theme() + theme(aspect.ratio = 1)
cluster_plot + genotype_plot
```

```{r}
markers <- FindMarkers(cxcl12_ko_meso_prelim, 18, NULL)
markers$gene <- rownames(markers)
markers
```

```{r}
immune_marker_genes <- c("Cd79a", "Cd19", "Ms4a1", # B
                         "Zbtb32", #Zbtb32+ B
                         "Cd27", "Slamf7", # Plasma
                         "Cd3e", "Lck", "Cd8a", # Cd8+ T
                         "Cd4", # Cd4+ T
                         "Ctla4", # Regulatory T
                         "Ly6g5b", # Ly6g5b+ T
                         "Alox5", # Alox5+ Lymphocyte
                         "Nkg7", "Cxcr6", # NKT
                         "Klrb1a", "Tyrobp", "Gzma", # NK
                         "Mki67", "Top2a", "Cdk1", # Prolif NK/T
                         "Retnlg", "S100a8", "S100a9", # Neutrophil
                         "Cpa3", "Ms4a2", "Mcpt8", # Basophil
                         "Marco", "Msr1", "Mcr1", # AM
                         "Siglech", "Cd300c", "Klk1", # Plasmacytoid Dendritic
                         "H2-Aa", "Cd24a", "Clec9a", # Myeloid Dendritic Type I
                         "Sirpa", "Itgam", # Myeloid Dendritic Type II
                         "Ccr7", # Ccr7+ Dendritic
                         "C1aq", "Trem2", # Interstitial macrophage
                         "Cd14", "F13a1", "Ly6c2", # Classical Monocyte
                         "Emr4", "Itgax", "Treml4" # Nonclassical Monocyte
                         # Intermediate Monocyte = Cd14+ Ly6c2-
)

immune_markers_reduced <- c("Ear1", # AM
                     "Plac8", # Monocyte
                     "H2-Aa", # CD
                     "C1qb" # IM
)

DotPlot(cxcl12_ko_immune_prelim, features = immune_markers_reduced) + theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

```{r}
FeaturePlot(cxcl12_ko_immune_prelim, features = c("Cd74", "H2-Aa", "Top2a", "Fn1"))
FeaturePlot(cxcl12_ko_immune_prelim, features = c("Ear1", "Plac8", "H2-Aa", "C1qb"))
```

```{r}
# No B and T cells seen, no Neutrophils seen
cxcl12_ko_immune_prelim_relabel <- cxcl12_ko_immune_prelim
cxcl12_ko_immune_prelim_relabel$celltype <- as.factor(Idents(cxcl12_ko_immune_prelim_relabel))

idxes <- levels(cxcl12_ko_immune_prelim_relabel$celltype)
labels <- rep("", length(levels(cxcl12_ko_immune_prelim_relabel$celltype)))
labels[idxes %in% c(3, 4)] <- "Interstitial Macrophage"
labels[idxes %in% c(0, 1, 2, 5, 6, 7)] <- "Alveolar Macrophage"
labels[idxes %in% c(8)] <- "Monocyte"
labels[idxes %in% c(9)] <- "Dendritic Cell"
labels

levels(cxcl12_ko_immune_prelim_relabel$celltype) <- labels

```

```{r}

cluster_plot <- DimPlot(cxcl12_ko_immune_prelim_relabel, label = TRUE) + umap_theme() + theme(aspect.ratio = 1)
celltype_plot <- DimPlot(cxcl12_ko_immune_prelim_relabel, group.by = "celltype") + umap_theme() + theme(aspect.ratio = 1)
genotype_plot <- DimPlot(cxcl12_ko_immune_prelim_relabel, group.by = "orig.ident") + umap_theme() + theme(aspect.ratio = 1)
cluster_plot + celltype_plot + genotype_plot
```

```{r}
immune_cluster_plot <- ggplot() + umap_theme() +
        ggrastr::rasterise(geom_point(data = cxcl12_ko_immune_prelim_relabel, aes(x = x, y = y, color = celltype), size = .05), dpi = 600) +
        scale_colour_manual(name = "Celltype", values = color_category_20,
                            guide = guide_legend(override.aes = list(size=3), order = 1)) +
        theme(aspect.ratio=1,
              legend.text=element_text(size=12),
              legend.title=element_text(size=12),
              panel.background = element_rect(fill = "transparent", colour = "black"),
              plot.background = element_rect(color = "transparent", fill = "transparent")) +
        ylab("UMAP2") + xlab("UMAP1")

immune_genotype_plot <- ggplot() + umap_theme() +
        ggrastr::rasterise(geom_point(data = cxcl12_ko_immune_prelim_relabel, aes(x = x, y = y, color = orig.ident), size = .05), dpi = 600) +
        scale_colour_manual(name = "Genotype", values = color_scanpy_default,
                            guide = guide_legend(override.aes = list(size=3), order = 1)) +
        theme(aspect.ratio=1,
              legend.text=element_text(size=12),
              legend.title=element_text(size=12),
              panel.background = element_rect(fill = "transparent", colour = "black"),
              plot.background = element_rect(color = "transparent", fill = "transparent")) +
        ylab("UMAP2") + xlab("UMAP1")

ggsave("./data/figures/cxcl12_prelim_umap_immune.pdf", immune_cluster_plot + immune_genotype_plot + plot_layout(ncol = 2) &
        plot_annotation(theme = theme(plot.background = element_rect(color = "transparent", fill = "transparent"))),
       bg = "transparent")

```

```{r}
saveRDS(cxcl12_ko_epi_prelim_relabel, "./data/20211123_cxcl12_ko_epithelium_labeled.rds", compress = "bzip2")
saveRDS(cxcl12_ko_endo_prelim_relabel, "./data/20211123_cxcl12_ko_endothelium_labeled.rds", compress = "bzip2")
saveRDS(cxcl12_ko_meso_prelim_relabel, "./data/20211123_cxcl12_ko_mesenchyme_labeled.rds", compress = "bzip2")
saveRDS(cxcl12_ko_immune_prelim_relabel, "./data/20211123_cxcl12_ko_immune_labeled.rds", compress = "bzip2")

```
